requirements.txt
numpy
scipy
torchvision>=0.1.8
matplotlib
tqdm
PyYAML
scikit-image
visdom1 >= 1.0.1
gdown